* Added SubmeshComponent base class and a handful of concrete components.
* Now components are fully integrated.Only remains a bug in lighting.
* Some speed improvements, by reducing shadow maps size, and bloom's textures size.
* Fixed non directional lights (their cube box was not created when not associated to a shadow map).
* Modified main debug panel overlays.
* Added IndexMapping class, to remove faces from Submesh.
* Fixed overlay rendering for rtot.
* Added LinesMapping index mapping class.
* Moved GeometryBuffers, fro Submesh and BillboardBase to RenderPipeline.
* CubeBoxManager now uses LinesMapping to generate its mesh.
* Updated importers.
* Removed built-in importers, except for OBJ and PLY

+++ source/Core/Castor3D/Src/Overlay/TextOverlay.hpp
@@ -32,7 +32,7 @@ namespace castor3d
 		*/
 		struct Vertex
 		{
-			int32_t coords[2];
+			float coords[2];
 			float text[2];
 			float texture[2];
 		};
